Extract MailViewerHeader
